tests/wpt/tests/trusted-types/TrustedTypePolicyFactory-createPolicy-cspTests-none-skip.html

<!DOCTYPE html>
<script src="/resources/testharness.js" ></script>
<script src="/resources/testharnessreport.js"></script>
<script src="support/helper.sub.js"></script>
<meta http-equiv="Content-Security-Policy" content="trusted-types * 'none' 'allow-duplicates'">
<body>
<script>
  test(t => {
    window.trustedTypes.createPolicy('SomeName', { createHTML: s => s } );
  }, "Can create policy with name 'SomeName'");

  test(t => {
    window.trustedTypes.createPolicy('SomeName', { createHTML: s => s } );
  }, "Can create a second policy with name 'SomeName'");

  test(t => {
    window.trustedTypes.createPolicy('default', { createHTML: s => s } );
  }, "Can create policy with name 'default'");
</script>
