LayoutTests/imported/w3c/web-platform-tests/trusted-types/TrustedTypePolicyFactory-createPolicy-cspTests-none-skip-expected.txt

CONSOLE MESSAGE: Unrecognized Content-Security-Policy directive 'trusted-types'.


FAIL Can create policy with name 'SomeName' undefined is not an object (evaluating 'window.trustedTypes.createPolicy')
FAIL Can create a second policy with name 'SomeName' undefined is not an object (evaluating 'window.trustedTypes.createPolicy')
FAIL Can create policy with name 'default' undefined is not an object (evaluating 'window.trustedTypes.createPolicy')
